Replace counters and Gauges with Prometheus
